refactor: #3-verify code값 없는 경우 수정 및 오류 추가
