+++ src/main/java/freshtrash/freshtrashbackend/exception/constants/ErrorCode.java
@@ -19,7 +19,8 @@ public enum ErrorCode {
     NOT_FOUND_WASTE(HttpStatus.NOT_FOUND, "폐기물 정보가 존재하지 않습니다."),
 
     // Mail
-    AUTH_CODE_UNMATCHED(HttpStatus.BAD_REQUEST, "잘못된 인증코드입니다.");
+    AUTH_CODE_UNMATCHED(HttpStatus.BAD_REQUEST, "잘못된 인증코드입니다."),
+    EMPTY_AUTH_CODE(HttpStatus.BAD_REQUEST, "인증코드가 입력되지 않았습니다.");
 
     private final HttpStatus status;
     private final String message;
